requirements.txt
python-binance
TA-Lib
numpy
pandas
matplotlib
websocket-client
